requirements.txt
python-telegram-bot==21.10
youtube-transcript-api==0.6.3
google-generativeai==0.8.4
python-dotenv==1.0.1
pytubefix==8.12.0
